Starting to parse schedules.
Firstly, we are now parsing LO and LI records, detailing the beginning
and intermediate stations of a schedule.

Secondly, this is generating a vast amount of data, so we have
switched to an 'online' approach, where functions are called to store
it away. We expect these to be written in subclasses.